Valoress

+++ src/main/java/com/projetoPOO/restaurante/model/itemCardapio/DadosCadastroItemCardapio.java
@@ -1,5 +1,7 @@
 package com.projetoPOO.restaurante.model.itemCardapio;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 
@@ -11,7 +13,7 @@ public record DadosCadastroItemCardapio(
     String componentes,
 
     @DecimalMin("0.01")
-    double preco
+    BigDecimal preco
 ) {
     public DadosCadastroItemCardapio(ItemCardapio itemCardapio) {
         this(itemCardapio.getNome(), itemCardapio.getComponentes(), itemCardapio.getPreco());
