+++ src/main/java/com/projetoPOO/restaurante/model/itemCardapio/DadosDetalhamentoItemCardapio.java
@@ -1,10 +1,12 @@
 package com.projetoPOO.restaurante.model.itemCardapio;
 
+import java.math.BigDecimal;
+
 public record DadosDetalhamentoItemCardapio(
     Long id,
     String nome,
     String componentes,
-    double preco
+    BigDecimal preco
 ) {
 
     public DadosDetalhamentoItemCardapio(ItemCardapio itemCardapio) {
